requirements.txt
telethon
selenium
python-dotenv